Fix touch continuity by using UITouch address as system ID (UP event coords may shift on iOS)

+++ include/reflex/pointer.h
@@ -21,7 +21,7 @@ namespace Reflex
 
 		public:
 
-			typedef int ID;
+			typedef uintptr_t ID;
 
 			enum Type
 			{
